deploy/requirements.txt
Flask
gunicorn
tensorflow==2.14.0
numpy
pillow